ARM: dts: bcm2711: Fix DMA constrains for newer BCM2711 boards
The commit 3d2cbb644836 ("ARM: dts: bcm2711: Move emmc2 into its own bus")
assumed that all bootloader pass the FDT modified by the firmware to the
kernel. But this is not always the case and cause boot issues for boards
with a BCM2711 C0 SoC (RPi 400, CM4, newer RPi 4 B) which does have
different DMA constraints:

  mmc0: error -22 whilst initialising SD card
  mmc0: invalid bus width

Since we are not able to detect the SoC revision, let's assume a BCM2711
C0 SoC for bcm2711.dtsi and move the restricted DMA constrains to the
RPi 4 B board file. Since the RPi 400 include the latter one we need
to overwrite the constraints again.

Fixes: 1c701accecf2 ("ARM: dts: Add Raspberry Pi 400 support")
Signed-off-by: Stefan Wahren <stefan.wahren@i2se.com>